+++ src/utils/utils.ts
@@ -1,0 +1,16 @@
+import { Experiment } from "../api/types"
+import { ExperimentDetailFieldsDescription } from "./constants"
+
+function isExperimentKeyTypeGuard(
+  key: string | keyof Experiment
+): key is keyof Experiment {
+  return key in ExperimentDetailFieldsDescription
+}
+
+export function getExperimentDetailFieldPlaceholder(
+  key: string | keyof Experiment
+) {
+  return isExperimentKeyTypeGuard(key)
+    ? ExperimentDetailFieldsDescription[key]
+    : key
+}
